Greatly simplified curvelet association
This simplifies code and makes it possible to debug. Though the
algorithm may not be quite as sophisticated, it works now, whereas
before it didn't work right for unknown reasons. The curvelets were not
being associated with the correct boundary segments. Likely due to the
low res boundary segment reconstruction. I fixed that in this build too.