Use second finest scale
My last checkin, I accidentally left my experimental change for the
third finest scale. This change can be useful in the situation where
there are high frequency noise bands in the FFT. The third finest scale
helps to eliminate this issue. It might be helpful to include the third
finest scale into the curvelet output. Also, we may want to normalize
the curvelet wedge before thresholding curvelets. This will also help
with noise artifacts that are only present in certain wedges.