Attempt to make this function more efficient
Now only draw line through center that is 2X dist threshold long. This
is not quite exact. If the fiber is on a 45, then the line will be sqrt2
longer than if it is horizontal or vertical. Should fix this, but not
sure how much it matters. Also, this is still not fast enough in my
opinion.