Clearer sections and zero fibers fix
I have segmented the ctFIRE code into many new sections, which are
labeled in plain english to facilitate user comprehension.

I have also implemented a fix for the edge situation where there is a
non-blank image that passes through the thresholds, but that does not
end up having any fibers within it.