Edge case - Full image analysis works, but produces no files
This fix addresses the edge case of an image with no fibers, but enough
curvelets to have gotten through CT-Fire.

The fix allows the program to skip over the image.  This is helpful for
the main analysis, but is problematic for further ROI analysis.  Will
need to add either a way to create files, or a check in ROI analysis to
find a "don't analyze" file.