Add bin count customization to histogram
Added new uislider and two private functions -- sliderValueChangedFcn and closeScriptFcn. Both help with identifying whether a new figure should be created when a bin count is modified, as without them, a new figure would be made every time the bin count was changed.

Utilized public variables BinNumber and UIAxes to keep track of the status of whether a new figure should be made or not, as well as the bin count of the histogram.