Create example_SHG_psf_image_simulation.py
This example simulates SHG point spread function(psf), SHG imaging and deconvolution . It is created using Gemini. For the 3D SHG psf,  it implements both a vectorial model and  a Gaussian approximation model with the former being more computationally demanding. Part of this example should be able to be adapted to the fiber generator main program.